Add interactive background image movement to index.html
Integrated a piece of JavaScript code to the index.html which will animate the background image based on mouse movement. This enhancement increases the user interactivity and gives a dynamic feel to the website. Additionally, CSS transition property has been added to create a smooth effect while the background image is being transformed.